requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
Faker==16.6.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
